pyproject.toml
[tool.poetry]
name = "kawapack"
version = "0.13.2"
description = "Utility for extracting Arknights assets"
authors = ["astral4 <88992929+astral4@users.noreply.github.com>"]
license = "AGPL-3.0-or-later"


[tool.poetry.dependencies]
python = "^3.11"
UnityPy = "^1.9.24"
Pillow = "^9.4.0"
pycryptodome = "^3.17.0"
pymongo = "^4.3.3"
fsb5 = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
